Limit updating of mode-line
Merely loading the library resulted in `keycast-mode-line-update'
being added to `pre-command-hook', which was clearly unnecessary.
Now the mode `keycast-mode' controls whether the hook function is
used.

Additionally only update the mode-line of the current buffer, by
calling `forge-mode-line-update' without the ALL argument.

The purpose of these changes isn't merely to reduce waste.  Forcing
all mode-lines to be updated causes `old-selected-window' to return
a different value, which causes issues for moody when the user moves
away from the still active minibuffer.