Restore updating while in minibuffer
In [1: 44f53d7] we started only updating the current buffer's
mode-line, among other things.  This not only broke updating
the mode-line when in the minibuffer, the given reason was also
wrong: the mention moody issue is not actually affected by this.

For a tiny performance boost, we keep limiting updating to the
current buffer if it isn't the minibuffer.

Closes #15.

1: 44f53d7cc62342d48fcba4d555ab204c8d7e482a
   Limit updating of mode-line