Fix keycast-log-mode breakage when reading a password
When a password is being read then `keycast--format' returns nil,
which is the right thing to do for `keycast-mode'.  The returned value
is used in the mode-line and if it is nil, that causes nothing to
appear in the mode-line, as opposed to the last binding from before
the password reading started to stick around.

However in the case of `keycast-log-mode' the value is being inserted
next to the previously inserted values using `insert' and that errors
on nil.  So for that mode `keycast--reading-password' has to be taken
into account elsewhere.

Fixes #20.