Use post-command-hook instead of pre-command-hook
A command may itself change the values of the relevant variables and
if it does it probably does so for a reason that we should respect.

We didn't do that before because `post-command-hook' comes with some
complications that force us to also use `minibuffer-exit-hook' for
some additional coordination.

An advantage of using `pre-command-hook' was that for a command that
takes a while to run that allowed displaying that command while it is
running, making it possible to see what is taking so long.