keycast-substitute-alist: Don't log transient-update
This command is invoked when invoking a prefix *key* (such as "-")
in a transient in order to re-display the transient buffer to dim
all suffixes whose bindings do not begin with that prefix.  The key
is then unread so that the user can continue typing the key sequence
they intended to type (such as "- f" to toggle the "-f" argument).

This command is an implementation detail that users don't need to
know about.